Fixup expectErr implementation
The Rust `expect_err` method on which this is based has the type
`fn expect_err(self,msg: &str) -> E` meaning that it returns the
container `Error` type not the `Ok` type. This change corrects the type
signature on BaseResult so the package builds.

Also adds a missing semicolon in `ErrImpl`.